Proper support for starting Resque
Turns out Scheduler *doesn't* start Resque; it just relies on Resque's code to avoid duplication of as much as possible.

This commit allows for proper support for starting both parts of the engine, with 5 actual workers and one scheduler.

Still not really any additional functionality, yet.

Did switch to a fork of resque-scheduler which supports the Credis-backed version of Resque.
May require slight patching of the currently-pulled-in version of Credis, especially without phpredis installed.

Signed-off-by: Daniel Hunsaker <danhunsaker@gmail.com>